feat: refactor tool usage in MessageContent and update imports for buildSystemPrompt
