+++ tsconfig.node.json
@@ -7,7 +7,7 @@
     "src/main/env.d.ts",
     "src/renderer/src/types/*",
     "packages/shared/**/*"
-, "src/renderer/src/services/mcp"  ],
+ ],
   "compilerOptions": {
     "composite": true,
     "types": [
